ref: moved merkle tree code into state module
And moved merkle tree storage into storage module.

This is in preparation for adding contract and global tree entry points,
instead of using the raw merkle tree.
