+++ crates/pathfinder/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod config;
 pub mod ethereum;
-pub mod merkle_tree;
 pub mod rpc;
 pub mod sequencer;
 pub mod serde;
+pub mod state;
 pub mod storage;
